+++ apps/server/src/schema/index.ts
@@ -3,7 +3,7 @@ import './errors';
 import './integrations/integration-operations';
 import './integrations/insights-operations';
 import './organization/org-types';
-import './user/organization-operations';
+import './organization/organization-operations';
 import './user/social-login-operations';
 import './user/user-operations';
